Move the get_hosts() method from the Stats_Calculator to the Attendee_Repository
