+++ includes/routes/event/details.php
@@ -59,7 +59,7 @@ class Details_Route extends Route {
 			$event_stats  = $stats_calculator->for_event( $event->id() );
 			$contributors = $stats_calculator->get_contributors( $event->id() );
 			$attendees    = $stats_calculator->get_attendees_not_contributing( $event->id() );
-			$hosts        = $stats_calculator->get_hosts( $event->id() );
+			$hosts        = $this->attendee_repository->get_hosts( $event->id() );
 			$projects     = $stats_calculator->get_projects( $event->id() );
 		} catch ( Exception $e ) {
 			// phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
